Fix tests to use moved api-client
