--- test/conftest.js
@@ -1,5 +1,0 @@
-"use strict";
-const chai = require('chai');
-const chaiAsPromised = require('chai-as-promised');
-
-chai.use(chaiAsPromised);
